refactor(dal): rename lib/api to lib/dal and add server-only
Rename the data access directory to lib/dal in preparation for adding
an explicit DAL/DTO boundary, and add 'server-only' to every file so
accidental client imports fail at build time. Update all import paths
across app, components, and actions accordingly. No runtime behavior
change.
