+++ src/lib/dal/client.ts
@@ -1,3 +1,5 @@
+import 'server-only';
+
 import { auth } from '@/lib/auth';
 
 export class ApiError extends Error {
