+++ src/lib/dal/examples.ts
@@ -1,3 +1,5 @@
+import 'server-only';
+
 import type {
   CreateExampleInput,
   Example,
