+++ src/lib/dal/meanings.ts
@@ -1,3 +1,5 @@
+import 'server-only';
+
 import type {
   CreateMeaningInput,
   Meaning,
