+++ src/lib/dal/settings.ts
@@ -1,3 +1,5 @@
+import 'server-only';
+
 import type {
   CreateSettingInput,
   Setting,
