+++ src/lib/dal/wordbooks.ts
@@ -1,3 +1,5 @@
+import 'server-only';
+
 import type {
   CreateWordbookInput,
   PaginatedResponse,
