+++ src/lib/dal/words.ts
@@ -1,3 +1,5 @@
+import 'server-only';
+
 import type {
   CreateWordInput,
   PaginatedResponse,
